+++ data/src/main/java/com/tht/tht/data/di/DataSourceModule.kt
@@ -6,16 +6,16 @@ import com.tht.tht.data.local.datasource.TermsDataSource
 import com.tht.tht.data.local.datasource.TermsDataSourceImpl
 import com.tht.tht.data.local.datasource.TokenDataSource
 import com.tht.tht.data.local.datasource.TokenDataSourceImpl
-import com.tht.tht.data.remote.datasource.EmailDataSource
-import com.tht.tht.data.remote.datasource.EmailDataSourceImpl
-import com.tht.tht.data.remote.datasource.ImageDataSource
-import com.tht.tht.data.remote.datasource.ImageDataSourceImpl
-import com.tht.tht.data.remote.datasource.LocationDataSource
-import com.tht.tht.data.remote.datasource.LocationDataSourceImpl
-import com.tht.tht.data.remote.datasource.RegionCodeDataSource
-import com.tht.tht.data.remote.datasource.RegionCodeDataSourceImpl
-import com.tht.tht.data.remote.datasource.SignupApiDataSource
-import com.tht.tht.data.remote.datasource.SignupApiDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.EmailDataSource
+import com.tht.tht.data.remote.datasource.signup.EmailDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.ImageDataSource
+import com.tht.tht.data.remote.datasource.signup.ImageDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.LocationDataSource
+import com.tht.tht.data.remote.datasource.signup.LocationDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.RegionCodeDataSource
+import com.tht.tht.data.remote.datasource.signup.RegionCodeDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.SignupApiDataSource
+import com.tht.tht.data.remote.datasource.signup.SignupApiDataSourceImpl
 import com.tht.tht.data.remote.datasource.dailyusercard.DailyUserCardDataSource
 import com.tht.tht.data.remote.datasource.dailyusercard.DailyUserCardDataSourceImpl
 import com.tht.tht.data.remote.datasource.login.LoginDataSource
