+++ data/src/main/java/com/tht/tht/data/remote/datasource/signup/ImageDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.tht.tht.data.remote.datasource
+package com.tht.tht.data.remote.datasource.signup
 
 import android.net.Uri
 import com.tht.tht.data.di.IODispatcher
