+++ data/src/main/java/com/tht/tht/data/repository/RegionCodeRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.tht.tht.data.repository
 
 import com.tht.tht.data.di.DefaultDispatcher
 import com.tht.tht.data.local.mapper.toModel
-import com.tht.tht.data.remote.datasource.RegionCodeDataSource
+import com.tht.tht.data.remote.datasource.signup.RegionCodeDataSource
 import com.tht.tht.domain.signup.model.RegionCodeModel
 import com.tht.tht.domain.signup.repository.RegionCodeRepository
 import kotlinx.coroutines.CoroutineDispatcher
