+++ data/src/test/java/com/tht/tht/data/datasource/LocationDataSourceImplTest.kt
@@ -1,6 +1,6 @@
 package com.tht.tht.data.datasource
 
-import com.tht.tht.data.remote.datasource.LocationDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.LocationDataSourceImpl
 import com.tht.tht.data.remote.response.location.LocationResponse
 import com.tht.tht.data.remote.service.location.LocationService
 import io.mockk.coEvery
