+++ data/src/test/java/com/tht/tht/data/datasource/RegionCodeDataSourceImplTest.kt
@@ -1,6 +1,6 @@
 package com.tht.tht.data.datasource
 
-import com.tht.tht.data.remote.datasource.RegionCodeDataSourceImpl
+import com.tht.tht.data.remote.datasource.signup.RegionCodeDataSourceImpl
 import com.tht.tht.data.remote.mapper.toUnwrap
 import com.tht.tht.data.remote.response.base.BaseResponse
 import com.tht.tht.data.remote.response.base.ErrorResponse
